Rename: home service를 client 폴더로 이동
